+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@ Release Notes
 # 0.2.0-SNAPSHOT (in progress)
 - default value, unit value and custom properties value are of type GenericValue now. This allows to be type safe and better supports JSON marshaling and unmarshalling.
 - Introduced GenericTypesystemContext to allow type and item accessibility checks based on owner and visibility attributes.
+- added method to create a new item with set default values for a given type
 - ...
 
 # 0.1.0-SNAPSHOT